+++ com/hbm/tileentity/TileEntityMachineShredder.java
@@ -148,7 +148,7 @@ public class TileEntityMachineShredder extends TileEntity implements ISidedInven
 		super.readFromNBT(nbt);
 		NBTTagList list = nbt.getTagList("items", 10);
 		
-		this.power = nbt.getShort("powerTime");
+		this.power = nbt.getInteger("powerTime");
 		slots = new ItemStack[getSizeInventory()];
 		
 		for(int i = 0; i < list.tagCount(); i++)
@@ -165,7 +165,7 @@ public class TileEntityMachineShredder extends TileEntity implements ISidedInven
 	@Override
 	public void writeToNBT(NBTTagCompound nbt) {
 		super.writeToNBT(nbt);
-		nbt.setShort("powerTime", (short) power);
+		nbt.setInteger("powerTime", power);
 		NBTTagList list = new NBTTagList();
 		
 		for(int i = 0; i < slots.length; i++)
